requirements.txt
requests
kaggle
sqlalchemy
fastapi
uvicorn
torch
torchvision
ultralytics
opencv-python-headless
numpy
pandas
scikit-learn
python-multipart
Pillow
